requirements.txt
aiogram>=3.4.1
aiohttp>=3.9.1
python-dotenv>=1.0.0
Django>=4.2.0
yandex-gpt>=0.1.0
